Adding basic error handling for failed API calls.
I periodically see urlopen throw `Connection timed out` errors when flickrapi attempts to connect to the Flickr API and get the photo pool. This causes Django to fail loading the relevant page with an Internal Server Error. To be slightly more graceful, the call to the API is now wrapped in a try/except clause. If the call fails, an empty pool will be returned.